rebrand to vaporplay
why?

Since the orignal name piongs was not ideal (like "a game server using
pion"), but it's too close to pion which makes it a bit wierd. So
before this repo goes public, I decided to rebrand it to vaporplay.
By no means this is an action to avoid the connection with the pion
community, pion is a great project created by amazing people, this
project couldn't possibly exist without the help of pion and it's
community.